Slim down cli plugin logic.
