+++ crates/nu-plugin/src/lib.rs
@@ -1,4 +1,6 @@
+pub mod jsonrpc;
 mod plugin;
+
 pub mod test_helpers;
 
 pub use crate::plugin::{serve_plugin, Plugin};
